Fix timeout error in BwcTest
Sometimes the `upgrade_segments` operation didn't work correctly.
`wait_for_active_shards` would timeout in `2.x` because a shard
contained old segment files that were no longer readable.

It seems that changing the number_of_replicas somehow caused this.
It's still unclear why - but since we cannot fix this back in 1.1 we'll
have to do a workaround here.

It also seems to be a race-condition, not something that is generally
broken.