Add Jenkinsfile
So that we can link the "how to run tests" with specific source
versions.

Otherwise if we want to extend the test suite old branches will fail
because the new test cases are not there yet.

This also makes use of the parallel pipeline feature to have a faster
feedback loop. If we're working on client tests we no longer need to
wait for the slow BWC tests to go.