Add aggregation stmt for rolling upgrade test
We had a regression on BWC for aggregations which have a different
intermediate input data type than the initial argument type causing
aggregation streamed from a 4.1 node to a 4.2 node to not being resolvable. 
This adds the related statement to the rolling upgrade test to ensure
that it will correctly work on mixed version clusters.

Relates https://github.com/crate/crate/pull/10557.