Fix flaky rolling upgrade update_docs recovery test
We must ensure that all replica shards are active before any node
is upgraded.
Otherwise the cluster may try to allocate new replica shards if
the upgraded node contains the primary shard. This would fail then
due to node version based allocation rules (a replica cannot be
allocated on a node which has a lower version than the primary).