Set netty worker_count for BWC upgrade tests
The tests are currently flaky, occassionally failing in 4.4. They get
stuck at the blob upload step and eventually timeout.

In 4.4.0 we made a change to share the event-loop group across transport
& http (https://github.com/crate/crate/commit/2df19cba5acf7444dbbad5eb0e88e4d13bc49b73)
which we later reverted in 4.7.0 because it can cause deadlocks: https://github.com/crate/crate/commit/38aeb09d3080916db716c08da94bc2310f003990

I assume the tests are flaky because of this deadlock.
This increases the netty worker count as a workaround.