Set worker count for rolling upgrade tests
The rolling upgrade tests are also flaky, sometimes getting stuck with a
`select count(*) from sys.shards where state = 'STARTED'` on the 4.4.0
upgrade step.

This change is mostly to rule out the threading deadlock mentioned in
https://github.com/crate/crate-qa/commit/cc3efd5ec3c24bc0fb0f7906dabc6f27462caaa1
as a reason. There are no blobs involved in the rolling upgrade, so this
likely doesn't fix it.