pgjdbc: Adjust test for updates to `conn.getMetaData().getCatalogs()`
As before, the inquiry of the available catalogs returned `doc` on the
first slot.

It looks actually correct now in 42.7, as `doc` is the schema, but not
the catalog. The SQL statement to inquire the available catalogs returns
`crate`, not `doc`.

  SELECT DISTINCT table_catalog FROM information_schema.tables 

This item on the upstream release notes seems to be relevant:

  feat: return all catalogs for getCatalogs metadata query closes
  ISSUE #2949 PR #2953

  -- https://jdbc.postgresql.org/changelogs/2023-11-20-42.7.0-release/