Changed upgrade paths, we don't support upgrades from 4.x to nightly
For a full cluster upgrade, we must fix the 4.x upgrade path to
end with any 5.x as tables created on 4.x won't be supported on
any 6.x version.

Rolling upgrades from 5.x to 6.x require a some fixes which are
still pending. Until this is merged, we must stop at the latest
5.x verison.