Add jdwp agent if running under debugpy
This adds logic to start the CrateDB instances with the jdwp agent,
listening on port '5{http.port}'.

This can be useful to debug a failure - e.g. you can set a breakpoint in
a crate-qa test case and then when it stops also attach to the CrateDB
instance and inspect the state there.